Final commit
This project was one of the easiest for me in this curriculum so far,
which I'm sure it wasn't meant to be incredibly difficult. It may have
also been easier because it was my first time using TDD :D. I ended up
writing 22 tests for the five functions we were tasked to create. I can
definitely see the value in using test driven development for projects.